requirements.txt
# ----------------------------
# Core API
# ----------------------------
fastapi==0.136.1
uvicorn==0.46.0
python-dotenv==1.2.2
httpx==0.28.1


# ----------------------------
# ML
# ----------------------------
numpy==2.4.4
scipy==1.17.1
pandas==3.0.2
scikit-learn==1.8.0
joblib==1.5.3

# ----------------------------
# Frontend
# ----------------------------
streamlit==1.57.0
